requirements.txt
# 核心框架
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
llama-index>=0.9.0

# 向量数据库
chromadb>=0.4.22
faiss-cpu>=1.7.4

# 文档解析
pypdf>=3.17.4
beautifulsoup4>=4.12.3
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2

# 文本处理
sentence-transformers>=2.3.1
tiktoken>=0.5.2
rank-bm25>=0.2.2

# LLM
openai>=1.10.0
cohere>=5.3.0
anthropic>=0.18.0

# 数据处理
numpy<2.0.0,>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Web服务
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# 评估
rouge-score>=0.1.2
bert-score

# 工具库
python-dotenv>=1.0.0
tqdm>=4.66.1
pydantic-settings>=2.1.0
